pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cellcounter"
dynamic = ["version"]
description = "Infers the number of cells in spatial spots from Visium spatial images."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
  { name = "Boshen Yan", email = "bosheny@andrew.cmu.edu" },
  { name = "Tanush Paradeshi", email = "tparades@andrew.cmu.edu" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10"
]
dependencies = [
  "anndata>=0.8.0",
  "scanpy>=1.9.0",
  "numpy>=1.20,<2.0",
  "pandas>=1.2",
  "matplotlib>=3.3",
  "Pillow>=8.0",
  "tensorflow>=2.5",
  "stardist>=0.8.0",
  "scikit-image>=0.23.0",
  "csbdeep>=0.8.0",
  "geopandas>=0.13.0",
  "shapely>=2.0.0",
  "rtree >= 0.3.0",
  "tifffile>=2024.8.30"
]

[project.optional-dependencies]
dev = [
  "black>=22.1.0",
  "pytest>=7.0"
]

[tool.setuptools.packages.find]
where = ["src"]
